+++ src/main/java/de/diddiz/LogBlock/Kill.java
@@ -30,7 +30,7 @@ public class Kill implements LookupCacheElement {
     }
 
     public Kill(ResultSet rs, QueryParams p) throws SQLException {
-        id = p.needId ? rs.getInt("id") : 0;
+        id = p.needId ? rs.getLong("id") : 0;
         date = p.needDate ? rs.getTimestamp("date").getTime() : 0;
         loc = p.needCoords ? new Location(p.world, rs.getInt("x"), rs.getInt("y"), rs.getInt("z")) : null;
         killerName = p.needKiller ? rs.getString("killer") : null;
